requirements.txt
opencv-python
pytesseract
openai
python-dotenv
Pillow
numpy
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
azure-iot-device
imap-tools
matplotlib